Move support/historyarchive to top level
